Fixed worng id check in display questionary
